refactored grid jsx components into separate files
