+++ src/App.jsx
@@ -1,10 +1,5 @@
-
 import React, { useState, useEffect } from 'react';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-} from 'react-router-dom';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 // component partials
 import Navbar from './components/Navbar/Navbar.jsx';
@@ -13,15 +8,16 @@ import Navbar from './components/Navbar/Navbar.jsx';
 import Login from './components/Login/LoginPage.jsx';
 import Register from './components/Register/RegisterPage.jsx';
 // main pages
-import GridElem from './components/grid.jsx';
+import GridElem from './components/World/Grid.jsx';
 // error pages
 import Error404 from './components/Error/Error404Page.jsx';
 
+import EditWorld from './components/World/EditWorld.jsx';
 
 function Grid() {
   return (
     <div className="pt-5">
-      <h1>Hello Jia En?</h1>
+      <h1>Hello Grid</h1>
       <GridElem />
     </div>
   );
@@ -33,6 +29,7 @@ export default function App() {
       <Navbar />
       <Switch>
         <Route exact path={['/', '/home', '/main']} component={Grid} />
+        <Route path="/edit" component={EditWorld} />
         <Route path="/register" component={Register} />
         <Route path="/login" component={Login} />
         <Route path="*" component={Error404} />
